also track example programs from tutorial
The only things I can't understand is "!" in draw-poligon definition (it
works without it, too!
But this repo is not about LISP ans its dialects (I don't know anything about it except
it's cool but cursed[1]). So far we are done with original tutorial and
can add small improvements on our own!

[1] http://www.winestockwebdesign.com/Essays/Lisp_Curse.html