small refactoring
split our tortoise-controlling functions in two parts -- basic ones, and
then use them to expose it to Guile.
The only problem is the tortoise_move function. C can not return several
values from a function easily. You might pass some out-parameters, but
this moves burden of parameters preparing to the calling site.
Another way is (and I going to do that) to declare new struct type
and use it as return parameter. The drawback is this struct will "pollute" all the code that
uses this function. Anonimous tuples are excellent for situation like
this, but it requires at least C++ (or, even better, Haskell or Rust --
tuples are first class citizens there!)