refactoring WIP -- make our code compile again
To resolve linking problem we first mark global_output as extern. It
basically says compler "do not create this variable, it will be
somewhere!". And indeed, we mention it again in main.c -- and
global_output in basics.c will refer to that in main.
Next, we remove definition of tortoise struct from header, and move
it to basics.c (it's used only by our functions, no other code depends
on it directly). Looks like we "hid" instance of tortoise struct from
other parts of code.